changed logic for category mapping
