+++ packages/ssr/src/component/article.js
@@ -57,8 +57,6 @@ module.exports = (client, analyticsStream, data, helmetContext) => {
             : providerData.article;
           const articleTemplate = article ? article.template : null;
 
-          const sectionPath = article ? getSectionNameForAnalytics(article) : "";
-
           return React.createElement(
             ContextProviderWithDefaults,
             {
@@ -80,9 +78,10 @@ module.exports = (client, analyticsStream, data, helmetContext) => {
               analyticsStream,
               article: {
                 ...article,
-                section: sectionPath || "unknown section",
-                isPreview,
-                url: article.url || articleDataFromRender.canonicalUrl || ""
+                section: article
+                  ? getSectionNameForAnalytics(article)
+                  : "unknown section",
+                isPreview
               },
               error,
               isLoading,
